frontend/pyproject.toml
[project]
name = "reflection"
version = "0.0.0"
requires-python = ">=3.13,<3.14"
dependencies = [
  "pydantic==2.12.5",
  "pydantic-settings==2.12.0",
  "pyyaml==6.0.3",
  "a2a-sdk[http-server]==0.3.24",
  "streamlit==1.56.0",
]


[project.optional-dependencies]
dev = [
  "ruff",
  "ipykernel",
  "pytest",
  "pytest-asyncio",
  "httpx",
]
